pyproject.toml
[tool.poetry]
name = "yana"
version = "0.1.0"
description = ""
authors = ["y.shurpik <y.shurpik@clearcode.cc>"]

[tool.poetry.dependencies]
python = "^3.10"
freezegun = "^1.2.1"
behave = "^1.2.6"
Faker = "^13.14.0"
black = "^22.6.0"
argparse = "^1.4.0"
pandas = "^1.4.3"
pytest-serverless = "^1.0.0"
tinydb = "^4.7.0"
pytest-freezegun = "^0.4.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
